Split request context id adding to separate middleware
The proxyHandler was adding a strictly increasing ID value to the
request's context that is used to correlate logs. As the only http
handler in the application, this was fine. However, moving to serving
PAC files from alpaca directly will split up the handling of http
requests over different handler for that and the proxy function. In this
case, it makes sense to split off the code that adds the ID to the
context into a separate piece of http middleware. This will ensure that
all requests have an ID, not just those handled by the proxy handler.

Add tests for the context ID.

Issues: #14