Make ProxyHandler HTTP middleware capable
Add a `WrapHandler()` method to `ProxyHandler` that can be used to wrap
another `http.Handler` such that any proxy requests (CONNECT methods or
absolute-form URIs) are passed to `ProxyHandler.ServeHTTP` and the rest
are passed to the next `http.Handler` (in the standard way that http
"middleware" works).

Plug in an empty `http.ServeMux` as the next handler. Any origin-form
URIs requested will go there and result in a 404 at the moment. This
will allow us to add paths to the mux in future commits so that alpaca
can serve content itself rather than just proxy requests. This will be
used to serve the PAC file that alpaca is using as well as serve a
wrapped PAC file that will direct all non-direct requests to alpaca.

Update the existing test cases to use the `WrapHandler()` method and
ensure the next `http.Handler` does not get called (an empty mux that
always returns 404).

Add an additional test to ensure that the middleware passes through
non-proxy requests (origin-form URIs) to the mux.

Move a few imports around as a result of `goimports` way of ordering
imports.

Issues: #14