Consolidate PAC URL, proxy finding and logging
Consolidate handing of the PAC URL into the `ProxyFinder` so that the
one function is used as the proxy function and all logging of DIRECT
proxy requests are in one place. This simplifies `main()` keeping the
proxy function out of it. It will also help when looking to consolidate
all request logging in one place.

Issues: #14