Fix lint errors
"Fix" lint errors by ignoring most of them - they are unchecked error
returns for the most part. The only thing that can really be done with
the errors would be to log them, but just add a nolint:errcheck comment
for now. They can be changed later to log if wanted.

One lint fix was a simple change for `a, _ := range` to `a := range`.

One lint fix was removing an unnecessary `var` declaration.

This change eliminates all lint errors for golangci-lint as of v1.17.1.