Introduce a `transport` type to write HTTP requests to a `net.Conn`
The purpose of this change is to reduce code duplication in
`authenticator.go`. Previously, we needed to have two separate
implementations of the NTLM handshake: one for CONNECT requests (which
are written directly to a `net.Conn`) and another for all other requests
(which are sent via an `http.Transport`). Now that we only need one
(written against any `http.RoundTripper`), it will be simpler to
implement additional authentication protocols in the future, such as
Kerberos.